Revert "[WIP] download page with download link when selection is full"
