+++ src/components/header.js
@@ -11,9 +11,6 @@ export class Header extends Component {
           </Link>
         </p>
         <p className="header-links">
-          <Link to="/download">
-            Download
-          </Link>
           <Link to="/doc">
             Documentation
           </Link>
