+++ src/redux/reducers/index.js
@@ -1,7 +1,7 @@
 import {combineReducers} from 'redux';
-import options from './options';
+import home from './home';
 import doc from './doc';
 import gulpAngular from './gulp-angular';
 import blog from './blog';
 
-export default combineReducers({options, doc, gulpAngular, blog});
+export default combineReducers({home, doc, gulpAngular, blog});
