+++ src/styles/home.scss
@@ -69,7 +69,8 @@
 
 .option-name {
   text-align: center;
-  font-size: 1.2rem;
+  font-family: $parisish;
+  font-size: 2rem;
   margin: 0;
 }
